Move tags from ApparentHorizons to GR Surfaces
