+++ src/PointwiseFunctions/GeneralRelativity/Surfaces/ComputeItems.hpp
@@ -27,8 +27,7 @@ class DataVector;
 #include "Utilities/SetNumberOfGridPoints.hpp"
 #include "Utilities/TMPL.hpp"
 
-namespace ah {
-namespace Tags {
+namespace StrahlkorperGr::Tags {
 /// @{
 /// These ComputeItems are different from those used in
 /// GeneralizedHarmonic evolution because these live only on the
@@ -95,5 +94,4 @@ struct SpatialChristoffelSecondKindCompute
   using base = ::gr::Tags::SpatialChristoffelSecondKind<DataVector, Dim, Frame>;
 };
 /// @}
-}  // namespace Tags
-}  // namespace ah
+}  // namespace StrahlkorperGr::Tags
