+++ tests/Unit/ApparentHorizons/CMakeLists.txt
@@ -7,7 +7,6 @@ set(LIBRARY_SOURCES
   Test_ApparentHorizonFinder.cpp
   Test_ComputeExcisionBoundaryVolumeQuantities.cpp
   Test_ComputeHorizonVolumeQuantities.cpp
-  Test_ComputeItems.cpp
   Test_FastFlow.cpp
   Test_InterpolationTargetApparentHorizon.cpp
   Test_ObserveCenters.cpp
