+++ tests/Unit/PointwiseFunctions/GeneralRelativity/Surfaces/CMakeLists.txt
@@ -4,7 +4,9 @@
 set(LIBRARY "Test_StrahlkorperGr")
 
 set(LIBRARY_SOURCES
+  Test_ComputeItems.cpp
   Test_StrahlkorperGr.cpp
+  Test_Tags.cpp
   )
 
 add_test_library(
